Grid
Made a grid of squares
Can hover over them and select them. They change colour when you do.

Buttons have a good system for this.
They need Interaction and Node
If I add Node to the sprites, it breaks something.
So I just added Interaction and handle it myself. It's a bit sloppy, but
it works.
fn interaction interaction::Hovered branch can't run
if it was hovered, it was set to none by the system that runs before it.
There's probably a better way to do this where I'm not just resetting all
interations to none every frame, but oh well.
I could add a Hovered Component and only reset those ones. Is that faster?
What is the cost of adding an empty component and querying it? Does it
even matter? probably not