Tower Menu and Building
Click any number of tiles then press the button in the UI to build towers.

UI
Made a side panel with 4 buttons
Each button makes a different tower

Tower
Make a tower at each selection
Towers are just different coloured squares atm.
They are a child of the Tile

No checks to see if you can spawn a tower, it just does it.
Doesn't care if there is already a tower there.