Floor, Numbers
Can turn the grid tiles into floor
Before you click on a minesweeper tile, I'm calling it a wall. Once it's
uncovered, it becomes the floor.
Floor is brown
Changed the green grass colour

Press f to make floor at a specific spot for testing

Numbers
Can set a number for a tile and then it will draw a number sprite on top
of itself.
Have a 9 sprite, but I don't really need it.
Check every frame to update numbers. Should just check when it's changed.
Used a sprite atlas so all the numbers can be in one image instead of 10
different pngs. Not sure if I did it quite right

Tile
get_colour
checks what state it's in - whether it is floor or wall. This lets the
tile store its own colour when it's made to make a checkerboard pattern.
It might've been better to have a global colour and have the tile just
store its evenness or calculate it.
hovering doesn't check its state, it just uses a hover colour

Grid
get_neighbours
8-connected
If on the edge, will return none

Tower
build towers before the selection is removed. If it's ambiguous, sometimes
it will get cleared before it is built.