Bombs
Press F to clear an area to turn it into floor
Randomly 'place' some bombs in the 5x5 ring around th 3x3 ring of floor.
Then for each edge tile of 3x3, calculate how many bombs it is touching

It doesn't actually place bombs, but picks some random positions among
those wall tiles bordering
Doesn't scale. Only works for the 3x3 set in code.
Can work for any 3x3, but might break if they overlap

Very complicated.
Turn center to floor
turn neighbours to floor
loop neighbours again to put the walls in a set
set is random so pull x random coords out of it
for each of these random coords, get its 3x3 neighbours. for each of thoes
that is a floor, add 1 to its number

TileInfo
tried to store Tile and Entity. It didn't work for me. Couldn't figure out
how to make it a reference and not a copy.

Selection Bug
Broke somehow
Click to select something
Click a button to build towers
Click away from the grid to deselect
I tried to make build_tower run before interaction, but it doesn't; they
run in a random order
changed to right click for deselect for now.