Battlement Style
Different way of making the territory grow.
Maximizes wall space to floor space
Kind of makes a battlement shape on a castle

Managed to implement this style to replace the old style in 60 lines
instead of 90. And it's easier to read and makes more sense. Fewer
redundant loops

One issue is the colouring of the floor and walls.
Might need to be swapped around to keep visual clarity. Stuff kinda blends
together.

Buttons
Show what colour of tower they will make.