Towers, Shooting
Swapped grass colours

Tower Component
Holds all the info a tower needs.
Made a field for most of the different option I want a tower to have.

Bullet Component
Tower has a bullet. When it wants to fire, create a sprite, add a copy of
the bullet onto it.

struct Movement
is a bit wonky. Feels like it's too deep. Tower.Bullet.Movement
Need to set what kind of movement it will have (shoot in this dir forever
or go to this point or follow this entity) but the tower doesn't know this
until it shoots.
Bullet.update_target does this when it gets spawned.
I might be cloning the bullet once more than I need to. Not sure.

Should the Structure be
Tower.Bullet.Movement
or Tower.Bullet and Tower.Movement?
Should movement be a component? Could add it to bullets and enemies.

TowerServer
Stores the different tower types
When you click a button, it inserts that tower.
Loosely coupled atm.
4 towers, 4 buttons, 4 colours.
If I increased any of them, it would probably break. The others wouldn't
adjust.
Need to fix that.
Sprite/colour should be attached to the tower somehow and should make the
right number of buttons based on the number of towers.

Towers
Basic - shoot towards stuff
shotgun - shoot multiple shots at once towards enemies
bomb - lob a bomb that then explodes in an area
swarm - spawn a numbre of bullets that float in the air for a sec before
seeking targets or a target point.
None of the special stuff is impl yet. They all just shoot 1 bullet when
you press T(for testing)

spawn_tower
Made a range indicator
Need to change to reflect the actual range and not just always 30
Need to be only visible when you click on that tower.
Alpha - when you stack multiple indicators, they overlap and become
opaque. Is there a way around that? Maybe a material or shader?
Line - can I just do a line indicator. Like a circle with no fill?