Tower Visuals, Shotgun
Linked tower visuals and buttons
The tower knows its colour and name.
The buttons reflect the names and the towers spawned use the set colour
If I made another tower, it should work fine, adding its own button with
a set name.

Shotgun
Towers can shoot a spread of bullets like a shotgun.
Can set the number in the spread and the cone angle.