Control Flow
Decided to try using a GameState to control what's going on at any given
time. Got the idea from the github template that doesn't work for itch
Start the game by loading the font and other stuff that needs loading

Then make the menu with a play button
The grid is also made here. It needs something loaded in the Loading state
so this was the next best. And it provides backdrop for the menu

Nothing happens until you hit play (you can't build towers and enemies
don't spawn). The side menus don't even appear

Then it switches to Gameplay
Spawing your castle is now auto. No more pressing F and C to start the
game.
Enemies start spawnign. I wanted to make it so the tutorial plays before
enemies start, but it's not in yet.

Tutorial state seems like more work. Do I have to add each system to both
Tutorial and Playing? Probably. Or some modified systems for tutorials.

This led to a lot of changes in the Plugin.build fns

Grid
Centered the grid in the x. Looks better a bit offset on y so the UI has
more space at the top of the screen

Fixed the Selection bug
Sometimes when you launched the game, you would click to build a tower and
the selection would dissappear before the spawn tower system ran that used
tiles with a selection component.
Fixed the ordering of systems.
Now you can click away from the grid to deselect or right click or press
space.

Main
Lib
Moved most stuff to lib. That's the way the other example from the other
github template has it.
Created a dark green background colour that looks better than grey.

UI
Main menu
Is just a play button
I could probably share code better here, but oh well.
switch_to_playing
There is a bug where sometimes you click through the button. You click the
button and it selects the tile underneath. I thought if I used an event to
offset the clicking of the button and changing states by 1 frame it would
fix it, but it doesn't. So this could be removed. Both clicking the button
and selecting a tile care about the frame the mouse is clicked so
offsetting that might be why it works sometimes.

spawn_director_menu
Left side menu
The idea is to show the enemy movements and thought processes. It's an AI
game jam so exposing this kinda helps the theme.
Doesn't do anything yet.