Clippy
Clearing up some of the warnings clippy gives me

Nothing seems to be broken.
It tried to fix a nested if, but it would've broke stuff.
Clippy doesn't notice that tick() does work so I want the first if to pass
first.
Unless it is smarter than me and it doesn't run the code after a && if the
first condition failed. So maybe it is safe as long as the tick is second.
No idea.