Version 0.2.0, Fixed Swarm Bug
Keep the version up to date with what builds I've released.

I think I was spawning the swarm helpers at the center so that's why an
enemy would be within range to spawn a bullet, but then not be in range
when it was time for the bullets to seek out an enemy.
Tower might be op nw that I increased its range.