Win and Lose
Can win or lose
To win, you need to fill the map and then survive 30s
To lose you need to get 0hp

Added a title to the start screen.

End game screen is pretty bad.
I'd need to do a fair bit of effort to make it able to reset the game. So
I just tell the player to hit F5.
The buttons are also ugly. I should've made a parent node so I could
center them. Spawning naked buttons kinda sucks.