SURVIVE
When in the endgame, the director UI just says SURVIVE

Fixed bug
At the strt, it would spawn 10 enemies. In the constructor, I was getting
10 posiitons. Changed it to 4. After the first spawn, it will calculate
batch size on its own like normal; it was just the first spawn that was
off.

Jam complete
This should be the last commit for the jam