Utils
Got better interception prediction
Uses math to figure out where the bullet will intercept the enemy
Added a crate to solve quadratics. I don't know if it is complicated to
solve quadratics in code or if there is some trick to solve it quick
and easy in code without sqrt or something.

Towers are still pretty dumb
They all ususally aim at the first enemy so when it dies, all the other
bullets that were midair will possibly miss, wasting a ton of DPS.

bezier curves
bezier curves are apparently really easy to make so I made some helpers
for them.
Not using them anywhere, but I could potentially use them for animation
curves and such. Maybe the movement of the swarm's first projectile.