Pathfinding
enemy
Changed<Enemy> on checking health so it only checks when it changes and
not every frame. Might be a small performance boost (probably negligible)

Lag
The real lag would be from the swarm towers. They spawn 5 things, The
bullets stay alive longer than other bullets because they have to get into
place before firing, and they have longer range so more towers are in range of
more enemies (probably the most impactful by far)
I didn't fix lag though.

Pathfinding
I wanted to implement A* in Rust
Used a tutorial by Code Monkey
It works, but it's not optimized really in any way.
The tutorial had some indicies backwards which was a nightmare to figure
out.
He creates the nodes in the order: 0, 4, 8, 12, 1, 5, ....
which works for him because he is setting their position in the array and
their value at the same time. array[0] = 0, array[4] = 4, array[8] = 8 ...
It didn't work when I tried to use a vec and
just push new values. My vec was ordered 0, 4, 8, 12, 1, 5, etc. so the
indicie and value are different, but they could just be the same. So the
code uses them interchangebly which makes it hard to figure out where it
is going wrong.
So I reversed the calculate_index to fix it.
left-to-right, down-to-up: x + y * width
down-to-up, left-to-right: y + x * width (should be height I guess)

Not implemented for anything, just as a test of A*