Flow Field
Different algorithm for pathfinding
Set a destination node and then update every tile to point towards the
path there.

Just implemented in the test file.

When an entity is over top of a tile, they will be able to reference its
direction and then use that direction to move towards the goal.

I should be able to run this once at the start to have every tile point
towards the castle.
I could then have towers with collision. When you place a tower (or wall
with collision), run this again to update and there will be a new flow
towards the castle that goes around walls.